Port to Qt 6 and drop Qt 5
libQGLViewer has no Qt 6 build - Debian and Ubuntu ship only
libqglviewer-dev-qt5 - so dropping Qt 5 meant replacing it. qgcoder used a
small corner of that library and already drew through vertex arrays, so
View is now a self-contained QOpenGLWidget on the OpenGL 3.3 core profile:
one shader, VAO/VBO geometry, and a QMatrix4x4 / QQuaternion camera with
orbit, pan and zoom. That drops libQGLViewer, GLEW, GLUT and GLU with it;
the build now needs only qt6-base-dev and libqgcodeeditor-qt6-dev.

The tool path is tessellated once into vertex buffers when it changes
instead of on every frame, straight moves use two samples and arcs are
sampled adaptively - so the old skip_dyn line skipping, which silently
dropped segments of large tool paths, is gone.

Elsewhere the Qt 6 idiom is applied throughout: qt_standard_project_setup()
and qt_add_executable, C++20, Qt::StringLiterals, std::unique_ptr for the Ui
object, QSettings by value, and pointer-to-member connects everywhere -
including the connections that used to live in mainwin.ui and settings.ui.
Qt 6.4 is the floor, which is what Ubuntu 24.04 LTS ships and the first
release with the _s and _L1 operators.

Fixes found along the way:

* a segfault on exit, from destroying the GL buffers both in ~View() and in
  the QOpenGLContext::aboutToBeDestroyed handler
* MainWindow::setStyle() and showFullScreen() shadowed non-virtual QWidget
  methods, as did View::update(); renamed to applyFontSize(),
  toggleFullScreen() and refresh()
* the command pane blocked the GUI thread in waitForFinished(). It is now an
  async QProcess fed on stdin, which also retires the fixed-path,
  world-executable /tmp/gcoder.sh it used to create and run
* changedGcode() called worker setters directly across a thread boundary
  instead of going through the queued signals
* the bounding box initialised its maxima to numeric_limits<double>::min(),
  the smallest positive double rather than the lowest one
* getSpindleMotionStatus() mixed two enum types in a bitwise or, deprecated
  in C++20

The render-rate readout moved from a GL overlay to the status bar: QPainter
on the QOpenGLWidget clobbered the shader's colour uniform.

+++ g2m/machineStatus.hpp
@@ -108,7 +108,7 @@ class machineStatus {
     /// return the current tool index
     int  getTool() const { return myTool; };
     /// return the current spindle status & motion type
-    int  getSpindleMotionStatus() const { return (spindleStat | motionType); }
+    int  getSpindleMotionStatus() const { return static_cast<int>(spindleStat) | static_cast<int>(motionType); }
 
   protected:
     /// machine Pose at start of this move
